refactor: Move the useActions hook to Hooks folder
