+++ src/components/Multiselect/MultiselectViewActions.jsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { useMultiSelection } from 'src/components/Contexts/MultiSelectionProvider'
+import { useActions } from 'src/components/Hooks/useActions'
 import MultiselectBackdrop from 'src/components/Multiselect/MultiselectBackdrop'
-import useActions from 'src/components/SearchResult/useActions'
 
 import ActionsBar from 'cozy-ui/transpiled/react/ActionsBar'
