+++ src/components/SearchResult/FlexsearchResultLine.jsx
@@ -2,8 +2,8 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import { useNavigate, useLocation } from 'react-router-dom'
 import { useMultiSelection } from 'src/components/Contexts/MultiSelectionProvider'
+import { useActions } from 'src/components/Hooks/useActions'
 import { makeFlexsearchResultLineOnClick } from 'src/components/SearchResult/helpers'
-import useActions from 'src/components/SearchResult/useActions'
 
 import { useClient } from 'cozy-client'
 import ListItemByDoc from 'cozy-ui/transpiled/react/ListItem/ListItemByDoc'
